+++ src/main/java/com/example/board/service/TaskService.java
@@ -8,12 +8,12 @@ import com.example.board.enity.Task;
 import java.util.List;
 
 public interface TaskService {
-    TaskResponse getTask(Long idTask);
+    TaskResponse getTask(Integer idTask);
     List<Task> getAllTask();
     TaskResponse createTask(TaskRequest taskRequest);
-    TaskResponse update(Long idTask, TaskRequest taskRequest);
-    TaskResponse statusUpdate(Long idTask, TaskRequest taskRequest);
+    TaskResponse update(Integer idTask, TaskRequest taskRequest);
+    TaskResponse statusUpdate(Integer idTask, TaskRequest taskRequest);
     //TaskResponse taskCount(Long idProject);
-    TaskResponse projectCompletion(Long idProject);
-    void delete(Long idTask);
+    TaskResponse projectCompletion(Integer idProject);
+    void delete(Integer idTask);
 }
